only trigger on first activated channel
Behaviour until now was to trigger on every channel independently.
This is not very useful in respect to phase-accuracy. Now we will
trigger on the first activated channel. Further development may
include a selectable trigger channel, but that requires GUI changes.

Additionally this fixes a thread-wait-for-itself issue while having no
active channels.