not stopping measure manually
QLenlab will not ask the controller to interrupt an measurement.
Instead, on the wirst call of controller::stop(), the thread will be
marked to stop after the next complete measurement. If you call
stop() a second time and it is not stopped yet, the measurement thread
will be killed after two seconds of patience.